requirements.txt
# Acquaplan Intelligent Tagger - Dependências Python

# Core dependencies
ollama>=0.1.0
Pillow>=10.0.0

# Google Drive (opcional - Projeto B)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Data processing
# (JSON já é built-in)

# Note: ExifTool deve ser instalado via sistema operacional
# macOS: brew install exiftool
# Linux: apt-get install libimage-exiftool-perl
